// client/cody/src/types.ts
export interface Message {
  speaker: 'you' | 'bot';
  text: string;
}

export interface ContextMessage extends Message {
  filename?: string;
}
// ...
export interface TranscriptChunk {
  actual: Message[];
  context: ContextMessage[];
  display?: Message[];
}
// ...
export interface Feedback {
  user: string;
  sentiment: 'good' | 'bad';
  displayMessages: Message[];
  transcript: TranscriptChunk[];
  feedbackVersion: string;
}